+++ PF-Client/package.json
@@ -17,9 +17,7 @@
     "react-bootstrap": "^2.6.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
-    "react-router-dom": "^6.4.3",
     "redux": "^4.2.0",
-    "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^2.4.2"
   },
   "devDependencies": {
