+++ PF-Client/src/App.css
@@ -1,5 +1,7 @@
 #root {
+  max-width: 1280px;
   margin: 0 auto;
+  padding: 2rem;
   text-align: center;
 }
