+++ PF-Client/src/index.css
@@ -5,8 +5,8 @@
   font-weight: 400;
 
   color-scheme: light dark;
-  color: rgba(0, 0, 0);
-  background-color: #f0f0f0;
+  color: rgba(255, 255, 255, 0.87);
+  background-color: #242424;
 
   font-synthesis: none;
   text-rendering: optimizeLegibility;
@@ -26,6 +26,7 @@ a:hover {
 
 body {
   margin: 0;
+  display: flex;
   place-items: center;
   min-width: 320px;
   min-height: 100vh;
