--- src/components/Nav/Nav.css
@@ -1,33 +1,0 @@
-.nnnn{
-  width: 100%;
-}
-
-.NavColorCss{
-    width: 99%;
-    font-size: 22px;
-    margin-top: 5px;
-    background-image: linear-gradient(to right, rgb(39, 7, 185) , rgb(133, 159, 236));
-    
-    color: aliceblue;
-    padding: 1.2% 0% 1.2% 0%;
-}
-
-.imgAvatar{
-    border: 2px solid rgb(230, 227, 227);
-    max-height: 42px;
-    
-}
-.logoNav{
-    font-weight: bolder;
-    cursor: pointer;
-}
-.colAvatarDrop{
-    width: 50%; 
- display: grid;
- grid-template-columns: 56% 5%;
- 
-}
-
-.dro{
-   margin-top: 6px;
-}
