--- src/main.jsx
@@ -1,18 +1,0 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import App from './App'
-import './index.css'
-import { Provider } from 'react-redux';
-import { BrowserRouter } from 'react-router-dom';
-import store from './redux/store/index';
-
-
-ReactDOM.createRoot(document.getElementById('root')).render(
-  <Provider store={store}>
-    <React.StrictMode>
-      <BrowserRouter>
-            <App />
-      </BrowserRouter>
-    </React.StrictMode>
-  </Provider>
-)
