--- src/redux/Reducer/index.js
@@ -1,18 +1,0 @@
-const initialState={
-    Profesores:[]
-}
-const rootReducer = (state = initialState, action) => {
-    switch (action.type) {
-     
-      case 'GET_PROFESORES':
-        return{
-            ...state
-        }
-        default:
-            return {
-              ...state
-            };
-        }
-}
-        
-      export default rootReducer;
