--- src/view/Home/Home.jsx
@@ -1,9 +1,0 @@
-import { NavBar } from "../../components/Nav/Nav"
-export const Home=()=>{
-    return(
-        <div>
-            <NavBar/>
-            Home.
-        </div>
-    )
-}
