--- src/view/Landing/Landing.jsx
@@ -1,7 +1,0 @@
-export const Landing =()=>{
-    return (
-        <div>
-            Landing
-        </div>
-    )
-}
