--- src/view/Perfil/Perfil.jsx
@@ -1,11 +1,0 @@
-import {PerfilProfesor} from '../PerfilProfesor/PerfilProfesor'
-import { PerfilAlumno } from '../PerfilAlumno/PerfilAlumno'
-
-export const Perfil=()=>{
-    const tipoUsuario='profesor'
-    return(
-        <>
-            {tipoUsuario==='profesor'?(<PerfilProfesor/>):(<PerfilAlumno/>)}
-        </>
-    )
-}
