--- src/view/PerfilAlumno/PerfilAlumno.jsx
@@ -1,7 +1,0 @@
-export const PerfilAlumno=()=>{
-    return(
-        <div>
-            Perfil Alumno
-        </div>
-    )
-}
