--- src/view/PerfilProfesor/PerfilProfesor.jsx
@@ -1,9 +1,0 @@
-import { NavBar } from "../../components/Nav/Nav"
-export const PerfilProfesor=()=>{
-    return(
-        <div>
-            <NavBar/>
-            Perfil Profesor
-        </div>
-    )
-}
